+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- renamed FileChecksum to FileHash
+
 ### Added
 
 - FileExists command `occ files:exists`